requirements.txt
fastapi
uvicorn
sqlmodel
starlette
python-multipart
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pyjwt
email-validator
psycopg2-binary
python-dotenv